+++ config/initializers/assets.rb
@@ -10,7 +10,6 @@ end
 Hourglass::Assets.precompile += %w(
   application.js application.css
   global.js global.css
-  swagger.js swagger.css swagger-print.css swagger-ui/images/*
   jqplot.js jqplot/jquery.jqplot.css
   icons/*.png
 )